move CappedToken.sol to ERC20Capped.sol, likewise for related files
