+++ contracts/mocks/RBACCappedTokenMock.sol
@@ -1,7 +1,7 @@
 pragma solidity ^0.4.24;
 
 import "../token/ERC20/RBACMintableToken.sol";
-import "../token/ERC20/CappedToken.sol";
+import "../token/ERC20/ERC20Capped.sol";
 
 
 contract RBACCappedTokenMock is ERC20Capped, RBACMintableToken {
