+++ test/token/ERC20/ERC20Capped.behavior.js
